fix: Add type assertion for getByteFrequencyData to resolve TypeScript error

+++ components/VoiceRecorder.tsx
@@ -49,7 +49,8 @@ export default function VoiceRecorder({ onRecordingComplete, onCancel, maxDurati
       const animate = () => {
         if (!analyserRef.current || !dataArrayRef.current) return;
         
-        analyserRef.current.getByteFrequencyData(dataArrayRef.current);
+        // Type assertion to fix TypeScript compatibility issue
+        analyserRef.current.getByteFrequencyData(dataArrayRef.current as Uint8Array);
         const newWaveform = Array.from(dataArrayRef.current.slice(0, 20)).map(value => value / 255);
         setWaveformData(newWaveform);
